# assignment8/backend/api/openapi.yaml
#
# CSE186 Assignment 8 Backend
#

openapi: '3.0.3'

info:
  description: CSE186 Assignment 8 Backend
  version: 4.1.0
  title: CSE186 Assignment 8 Backend

servers:
  - url: http://localhost:3010/api/v0

paths:
  /auth/login:
    post:
      summary: Log in with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
# Should be singular post
  /posts:
    get:
      summary: List posts visible to the authenticated user
      operationId: listPosts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /posts/{postId}/reactions:
    put:
      summary: Set or change the current user's reaction on a post
      operationId: setPostReaction
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
              properties:
                type:
                  $ref: '#/components/schemas/ReactionType'
      responses:
        '204':
          description: Reaction set
        '400':
          description: Invalid reaction type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Remove the current user's reaction from a post
      operationId: deletePostReaction
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: Reaction removed
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /groups:
    get:
      summary: List groups the authenticated user is a member of
      operationId: listGroups
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
#  /group/{id}/post:
  /groups/{groupId}/posts:
    get:
      summary: List posts in a specific group for the authenticated user
      operationId: listGroupPosts
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of posts in the group
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - user is not a member of the group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ReactionType:
      type: string
      enum:
        - like
        - love
        - haha
        - wow
        - sad
        - angry
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1

    User:
      type: object
      required:
        - id
        - email
        - displayName
      properties:
        id:
          type: integer
          format: int32
        email:
          type: string
          format: email
        displayName:
          type: string
        roles:
          type: array
          items:
            type: string

    LoginResponse:
      type: object
      required:
        - token
        - user
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'

    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: object
        status:
          type: integer

    Post:
      type: object
      required:
        - id
        - authorId
        - content
      properties:
        id:
          type: integer
          format: int32
        authorId:
          type: integer
          format: int32
        authorDisplayName:
          type: string
          description: Author display name for UI
        content:
          type: object
          required:
            - text
            - createdAt
          properties:
            text:
              type: string
            createdAt:
              type: string
              format: date-time
            visibility:
              type: string
              enum:
                - PUBLIC
                - FRIENDS
                - GROUP
            groupId:
              type: string
              format: uuid
        reactions:
          type: object
          additionalProperties:
            type: integer
            format: int32
          description: Aggregated reaction counts keyed by reaction type
        userReaction:
          $ref: '#/components/schemas/ReactionType'

    Group:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
 